backend/history-mastodon/requirements.txt
# COMP90024 Team 75 
# Linying Wei (1638206)
# Wangyang Wu (1641248)
# Ziyu Wang (1560831)
# Roger Zhang (1079986)

elasticsearch==8.17.1
elastic-transport<9,>=8.15.1
Flask==3.1.1
python_dateutil==2.9.0.post0
Requests==2.32.3
urllib3==2.4.0
beautifulsoup4==4.13.4
vaderSentiment==3.3.2